+++ app/helpers/api/bapi/getSolution.js
@@ -1,0 +1,11 @@
+import { getData } from 'buying-catalogue-library';
+import { logger } from '../../../logger';
+import { getEndpoint } from '../../../endpoints';
+
+export const getSolution = async ({ solutionId, accessToken }) => {
+  const endpoint = getEndpoint({ api: 'bapi', endpointLocator: 'getSolution', options: { solutionId } });
+  const solutionData = await getData({ endpoint, accessToken, logger });
+  logger.info(`Retrived solution data from BAPI for ${solutionId}`);
+
+  return solutionData;
+};
