+++ app/pages/sections/additional-services/select/recipient/contextCreator.js
@@ -2,9 +2,9 @@ import manifest from './manifest.json';
 import { baseUrl } from '../../../../../config';
 
 export const getContext = ({
-  orderId, solutionName,
+  orderId, itemName,
 }) => ({
   ...manifest,
-  title: `${manifest.title} ${solutionName}`,
+  title: `${manifest.title} ${itemName}`,
   backLinkHref: `${baseUrl}/organisation/${orderId}/additional-services/select/additional-service/price`,
 });
