+++ app/pages/sections/additional-services/select/recipient/controller.js
@@ -1,14 +1,3 @@
-import { getData } from 'buying-catalogue-library';
 import { getContext } from './contextCreator';
-import { getEndpoint } from '../../../../../endpoints';
-import { logger } from '../../../../../logger';
 
 export const getAdditionalServiceRecipientPageContext = params => getContext(params);
-
-export const getSolution = async ({ solutionId, accessToken }) => {
-  const endpoint = getEndpoint({ api: 'bapi', endpointLocator: 'getSolution', options: { solutionId } });
-  const solutionData = await getData({ endpoint, accessToken, logger });
-  logger.info(`Retrived solution data from BAPI for ${solutionId}`);
-
-  return solutionData;
-};
